+++ resources/js/components/StoryForm.vue
@@ -22,7 +22,7 @@
                     <div class="form-group">
                         <div class="custom-control custom-switch">
                             <input type="checkbox" class="custom-control-input" id="is_private" v-model="form.is_private" :disabled="loading">
-                            <label class="custom-control-label" for="is_private" dusk="link-form-private">{{ __('Private link?') }}</label>
+                            <label class="custom-control-label" for="is_private" dusk="story-form-private">{{ __('Private story?') }}</label>
                         </div>
                     </div>
                 </div>
@@ -31,7 +31,7 @@
                     <div class="form-group">
                         <div class="custom-control custom-switch">
                             <input type="checkbox" class="custom-control-input" id="is_pinned" v-model="form.is_pinned" :disabled="loading">
-                            <label class="custom-control-label" for="is_pinned" dusk="link-form-pinned">{{ __('Is pinned?') }}</label>
+                            <label class="custom-control-label" for="is_pinned" dusk="story-form-pinned">{{ __('Is pinned?') }}</label>
                         </div>
                     </div>
                 </div>