+++ app/Http/Controllers/LinkArchiveController.php
@@ -2,9 +2,7 @@
 
 namespace App\Http\Controllers;
 
-use App\Events\LinkArchiveRequested;
 use App\Link;
-use App\Services\LinkArchive\LinkArchive;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Str;
@@ -13,33 +11,7 @@ class LinkArchiveController extends Controller
 {
     public function __construct()
     {
-        $this->middleware('auth')
-            ->except('download');
-
-        $this->middleware('demo')
-            ->except('form');
-    }
-
-    public function form(Request $request, int $id)
-    {
-        /** @var Link $link */
-        $link = Link::findOrFail($id);
-
-        return view('link-archive')->with([
-            'page_title' => __('Manage archive'),
-            'providers' => LinkArchive::availableFor($link->url),
-            'link' => $link,
-        ]);
-    }
-
-    public function store(Request $request, int $id)
-    {
-        /** @var Link $link */
-        $link = Link::findOrFail($id);
-        event(new LinkArchiveRequested($link, $request->get('type')));
-
-        $this->flash(__('Link is being archived'), 'success');
-        return redirect()->away($link->permalink);
+        $this->middleware('demo');
     }
 
     public function download(Request $request, int $id, string $hash)
@@ -62,26 +34,6 @@ class LinkArchiveController extends Controller
         }
 
         $this->flash(__('Link archive doest not exist', 'error'));
-        return redirect()->back();
-    }
-
-    public function delete(Request $request, int $id)
-    {
-        /** @var Link $link */
-        $link = Link::findOrFail($id);
-
-        $file = $link->archive;
-
-        if ($file && Storage::disk('archives')->exists($file)) {
-            Storage::disk('archives')->delete($file);
-            $link->archive = null;
-            $link->save();
-
-            $this->flash(__('Archive has been deleted'), 'success');
-            return redirect()->back();
-        }
-
-        $this->flash(__('Archive doest not exist'), 'error');
         return redirect()->back();
     }
 }