+++ changelog.md
@@ -14,6 +14,7 @@
 
 - Translation in german updated (thanks to [kayschima](https://github.com/kayschima))
 - Select tags now use `custom-select` class
+- Move vue component styles to `app.scss`
 - Typos in StoryForm
 
 # 1.2.21